fix: focus lost on split
