+++ packages/slate/src/index.ts
@@ -1,4 +1,4 @@
-export * from './Slate';
+export * from './SlateEditor';
 export * from './contexts';
 export * from './caret';
 export * from './normalization';
